requirements.txt
httpx>=0.27
python-dotenv>=1.0
py-clob-client>=0.20
